@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: ">=3.0.5 <4.0.0"
 
 dependencies:
   flutter:
@@ -22,18 +22,27 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  animations: ^1.0.0+5
-  lipsum: ^1.0.1
-  font_awesome_flutter: ^8.7.0
-  provider: ^4.0.4
-  lottie: ^0.3.2
-  dio: ^3.0.9
-  socket_io_client: ^0.9.9
+  cupertino_icons: ^1.0.5
+  animations: ^2.0.7
+  lipsum:
+    git:
+      url: https://github.com/VB10/lipsum-dart
+  font_awesome_flutter: ^10.5.0
+  provider: ^6.0.5
+  lottie: ^2.4.0
+  dio: ^5.2.1+1
+  socket_io_client: ^2.0.2
+
+  json_annotation: ^4.8.1
+  equatable: ^2.0.5
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+  very_good_analysis: ^5.0.0+1
+  json_serializable: ^6.7.1
+  build_runner: ^2.4.6
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
