@@ -4,29 +4,29 @@ import 'package:shoppifront/features/product/model/product.dart';
 
 class GlobalModel {
   int index;
-  ProductModel model;
+  Product? model;
   GlobalModel({
-    this.index,
-    this.model,
+    required this.index,
+    required this.model,
   });
 
   Map<String, dynamic> toMap() {
     return {
       'index': index,
-      'model': model.toJson(),
+      'model': model?.toJson(),
     };
   }
 
-  static GlobalModel fromMap(Map<String, dynamic> map) {
+  static GlobalModel? fromMap(Map<String, dynamic>? map) {
     if (map == null) return null;
 
     return GlobalModel(
       index: map['index'],
-      model: ProductModel.fromJson(map['model']),
+      model: Product.fromJson(map['model']),
     );
   }
 
   String toJson() => json.encode(toMap());
 
-  static GlobalModel fromJson(String source) => fromMap(json.decode(source));
+  static GlobalModel? fromJson(String source) => fromMap(json.decode(source));
 }
