@@ -3,7 +3,7 @@
 import 'package:dio/dio.dart';
 
 class BaseConfig {
-  Dio dio; // with default Options
+  late final Dio dio; // with default Options
   final String baseUrl = "http://localhost:4000/";
 
   BaseConfig() {
@@ -12,13 +12,5 @@ class BaseConfig {
     );
 
     dio = Dio(options);
-    dio.interceptors
-        .add(InterceptorsWrapper(onRequest: (RequestOptions options) async {
-      return options; //continue
-    }, onResponse: (Response response) async {
-      return response; // continue
-    }, onError: (DioError e) async {
-      return e; //continue
-    }));
   }
 }
