@@ -1,16 +1,14 @@
 part of '../base_service.dart';
 
 class _RoutePath {
-  static _RoutePath _instance;
+  const _RoutePath._init();
+
+  static _RoutePath? _instance;
 
   static _RoutePath get instance {
-    if (_instance == null) {
-      _instance = _RoutePath._init();
-    }
-    return _instance;
+    _instance ??= _RoutePath._init();
+    return _instance!;
   }
-
-  _RoutePath._init();
 
   String get product => "product";
 }
