@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
 
 class NumberCircleAvatar extends StatelessWidget {
+  const NumberCircleAvatar({super.key, this.index = 0});
   final int index;
-
-  const NumberCircleAvatar({Key key, this.index = 0}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return CircleAvatar(
       child: Text(
-        index == 0 ? "" : index.toString(),
+        index == 0 ? '' : index.toString(),
         style: Theme.of(context)
             .textTheme
-            .headline5
-            .copyWith(fontWeight: FontWeight.bold),
+            .headlineSmall
+            ?.copyWith(fontWeight: FontWeight.bold),
       ),
     );
   }
