@@ -3,11 +3,9 @@ import 'package:shoppingapp/core/view/base/base_stateless.dart';
 import 'package:shoppingapp/features/shop/model/product.dart';
 
 class ShoppingCard extends BaseStatelessWidget {
+  const ShoppingCard({required this.product, super.key, this.padding});
   final Product product;
-  final EdgeInsets padding;
-  final double radius;
-
-  ShoppingCard({Key key, this.product, this.padding, this.radius});
+  final EdgeInsets? padding;
   @override
   Widget build(BuildContext context) {
     final currentTheme = Theme.of(context);
@@ -15,12 +13,13 @@ class ShoppingCard extends BaseStatelessWidget {
     return Card(
       elevation: 5,
       shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.all(
-              Radius.circular(dynamicWidth(context: context, val: 0.04)))),
+        borderRadius: BorderRadius.all(
+          Radius.circular(dynamicWidth(context: context, value: 0.04)),
+        ),
+      ),
       child: Padding(
-        padding: padding ?? EdgeInsets.all(10),
+        padding: padding ?? const EdgeInsets.all(10),
         child: Column(
-          mainAxisAlignment: MainAxisAlignment.start,
           crossAxisAlignment: CrossAxisAlignment.start,
           children: <Widget>[
             buildExpandedImage(),
@@ -37,33 +36,42 @@ class ShoppingCard extends BaseStatelessWidget {
 
   Text buildWeightText(ThemeData currentTheme) {
     return Text(
-      "${product.weight.toInt()} g",
-      style: currentTheme.textTheme.headline6
-          .copyWith(fontWeight: FontWeight.w300),
+      '${product.weight} g',
+      style: currentTheme.textTheme.titleLarge
+          ?.copyWith(fontWeight: FontWeight.w300),
     );
   }
 
   Text buildTitleText(ThemeData currentTheme) {
     return Text(
-      product.title,
+      product.title ?? '',
       maxLines: 2,
-      style: currentTheme.textTheme.headline6,
+      style: currentTheme.textTheme.titleLarge,
     );
   }
 
   Text buildText(ThemeData currentTheme) {
     return Text(
-      "\$ ${product.price}",
-      style: currentTheme.textTheme.headline4
-          .copyWith(fontWeight: FontWeight.bold),
+      '\$ ${product.price}',
+      style: currentTheme.textTheme.headlineMedium
+          ?.copyWith(fontWeight: FontWeight.bold),
     );
   }
 
   Expanded buildExpandedImage() {
+    if (product.image == null) {
+      return const Expanded(
+        child: Center(
+          child: Icon(Icons.error),
+        ),
+      );
+    }
     return Expanded(
       child: Center(
         child: AspectRatio(
-            aspectRatio: 1 / 3, child: Image.network(product.image)),
+          aspectRatio: 1 / 3,
+          child: Image.network(product.image!),
+        ),
       ),
     );
   }
