@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
-import '../../../constants/app_constants.dart';
-import '../../../constants/app_strings.dart';
+import 'package:shoppingapp/core/constants/app_constants.dart';
+import 'package:shoppingapp/core/constants/app_strings.dart';
 
 class PaymentListTile extends StatelessWidget {
+  const PaymentListTile({super.key, required this.totalMoney});
   final int totalMoney;
-
-  const PaymentListTile({Key key, @required this.totalMoney}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return ListTile(
@@ -17,24 +16,26 @@ class PaymentListTile extends StatelessWidget {
       subtitle: Row(
         children: <Widget>[
           Expanded(flex: 2, child: buildDeliveryBodyColumn(context)),
-          Spacer(),
+          const Spacer(),
         ],
       ),
     );
   }
 
   Text buildTextMoney(BuildContext context) {
-    return Text("\$${AppConstants.FREE_DELIVERY_MONEY}",
-        style: Theme.of(context)
-            .primaryTextTheme
-            .headline4
-            .copyWith(color: Theme.of(context).splashColor));
+    return Text(
+      '\$${AppConstants.FREE_DELIVERY_MONEY}',
+      style: Theme.of(context)
+          .primaryTextTheme
+          .headlineMedium
+          ?.copyWith(color: Theme.of(context).splashColor),
+    );
   }
 
   Text buildTextDelivery(BuildContext context) {
     return Text(
       AppStrings.instance.delivery,
-      style: Theme.of(context).primaryTextTheme.headline3,
+      style: Theme.of(context).primaryTextTheme.displaySmall,
     );
   }
 
@@ -51,7 +52,7 @@ class PaymentListTile extends StatelessWidget {
   Column buildDeliveryBodyColumn(BuildContext context) {
     return Column(
       children: <Widget>[
-        Spacer(),
+        const Spacer(),
         buildExpandedFreeText(context),
         buildLinearProgressIndicator(context),
       ],
@@ -74,11 +75,11 @@ class PaymentListTile extends StatelessWidget {
         AppStrings.instance.orderFree,
         style: Theme.of(context)
             .primaryTextTheme
-            .headline5
-            .copyWith(color: Theme.of(context).splashColor),
+            .headlineSmall
+            ?.copyWith(color: Theme.of(context).splashColor),
       ),
     );
   }
 
-  double get indicatorValue => (totalMoney / AppConstants.FREE_DELIVERY_MONEY);
+  double get indicatorValue => totalMoney / AppConstants.FREE_DELIVERY_MONEY;
 }
