@@ -2,24 +2,23 @@ import 'package:flutter/material.dart';
 import 'package:shoppingapp/core/view/base/base_state.dart';
 
 class NumberPicker extends StatefulWidget {
-  final int number;
-  final Function(int value) onChanged;
+  const NumberPicker({
+    required this.onChanged,
+    this.number,
+    super.key,
+    this.maxValue = 100,
+    this.minValue = 0,
+  });
+  final int? number;
+  final void Function(int value) onChanged;
   final int maxValue;
   final int minValue;
-
-  const NumberPicker(
-      {Key key,
-      @required this.number,
-      this.onChanged,
-      this.maxValue = 100,
-      this.minValue = 0})
-      : super(key: key);
   @override
   _NumberPickerState createState() => _NumberPickerState();
 }
 
 class _NumberPickerState extends BaseState<NumberPicker> {
-  int number;
+  late final int number;
 
   @override
   void initState() {
@@ -28,8 +27,10 @@ class _NumberPickerState extends BaseState<NumberPicker> {
   }
 
   ShapeDecoration get shapeDecoration => ShapeDecoration(
-      shape: StadiumBorder(
-          side: BorderSide(color: Theme.of(context).highlightColor)));
+        shape: StadiumBorder(
+          side: BorderSide(color: Theme.of(context).highlightColor),
+        ),
+      );
 
   @override
   Widget build(BuildContext context) {
@@ -40,10 +41,11 @@ class _NumberPickerState extends BaseState<NumberPicker> {
         children: <Widget>[
           deIncerement,
           Text(
-            "$number",
-            style: currentTheme.textTheme.headline4.copyWith(
-                fontWeight: FontWeight.bold,
-                color: currentTheme.colorScheme.onSurface),
+            '$number',
+            style: currentTheme.textTheme.headlineMedium?.copyWith(
+              fontWeight: FontWeight.bold,
+              color: currentTheme.colorScheme.onSurface,
+            ),
           ),
           incerement
         ],
@@ -52,18 +54,20 @@ class _NumberPickerState extends BaseState<NumberPicker> {
   }
 
   IconButton get deIncerement => IconButton(
-      icon: Icon(Icons.remove),
-      onPressed: () {
-        number -= number > widget.minValue ? 1 : 0;
-        updateNumber();
-      });
+        icon: const Icon(Icons.remove),
+        onPressed: () {
+          number -= number > widget.minValue ? 1 : 0;
+          updateNumber();
+        },
+      );
 
   IconButton get incerement => IconButton(
-      icon: Icon(Icons.add),
-      onPressed: () {
-        number += number <= widget.maxValue ? 1 : 0;
-        updateNumber();
-      });
+        icon: const Icon(Icons.add),
+        onPressed: () {
+          number += number <= widget.maxValue ? 1 : 0;
+          updateNumber();
+        },
+      );
 
   void updateNumber() {
     setState(() {});
