@@ -1,15 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:shoppingapp/core/constants/app_strings.dart';
+import 'package:shoppingapp/core/view/base/base_state.dart';
+import 'package:shoppingapp/core/view/widget/card/shop_payment_card.dart';
+import 'package:shoppingapp/core/view/widget/listTile/payment_list_tile.dart';
 import 'package:shoppingapp/core/view/widget/lottie/not_found.dart';
 import 'package:shoppingapp/features/notifier/product_list_notifier.dart';
-
-import '../../../core/constants/app_strings.dart';
-import '../../../core/view/base/base_state.dart';
-import '../../../core/view/widget/card/shop_payment_card.dart';
-import '../../../core/view/widget/listTile/payment_list_tile.dart';
-import '../model/product.dart';
+import 'package:shoppingapp/features/shop/model/product.dart';
 
 class ShopPaymentView extends StatefulWidget {
+  const ShopPaymentView({super.key});
+
   @override
   _ShopPaymentViewState createState() => _ShopPaymentViewState();
 }
@@ -24,16 +25,15 @@ class _ShopPaymentViewState extends BaseState<ShopPaymentView> {
     return Padding(
       padding: EdgeInsets.symmetric(horizontal: dynamicWidth(0.08)),
       child: Column(
-        mainAxisAlignment: MainAxisAlignment.start,
         crossAxisAlignment: CrossAxisAlignment.start,
         children: <Widget>[
-          Spacer(),
+          const Spacer(),
           cartTitle(),
           buildExpandedShopList(),
           buildDeliveryCard(),
-          Spacer(),
+          const Spacer(),
           buildMoneyTotalCard(context),
-          Spacer(),
+          const Spacer(),
           buildNextButton(context),
         ],
       ),
@@ -45,25 +45,31 @@ class _ShopPaymentViewState extends BaseState<ShopPaymentView> {
   Text cartTitle() {
     return Text(
       AppStrings.instance.paymentTitle,
-      style: currentTheme.primaryTextTheme.headline1,
+      style: currentTheme.primaryTextTheme.displayLarge,
       textScaleFactor: 1.5,
     );
   }
 
 // 2
   Expanded buildDeliveryCard() => Expanded(
-      flex: 1, child: PaymentListTile(totalMoney: sumMoneyAll.toInt()));
+        child: PaymentListTile(totalMoney: sumMoneyAll.toInt()),
+      );
 
   SizedBox buildNextButton(BuildContext context) {
     return SizedBox(
       width: dynamicWidth(1),
       height: dynamicHeight(0.07),
-      child: RaisedButton(
-          shape: StadiumBorder(),
-          color: Theme.of(context).primaryColorDark,
-          onPressed: () {},
-          child: Text(AppStrings.instance.next,
-              style: currentTheme.textTheme.headline3)),
+      child: ElevatedButton(
+        style: ElevatedButton.styleFrom(
+          shape: const StadiumBorder(),
+          backgroundColor: Theme.of(context).primaryColorDark,
+        ),
+        onPressed: () {},
+        child: Text(
+          AppStrings.instance.next,
+          style: currentTheme.textTheme.displaySmall,
+        ),
+      ),
     );
   }
 
@@ -75,34 +81,36 @@ class _ShopPaymentViewState extends BaseState<ShopPaymentView> {
             AppStrings.instance.total,
             style: Theme.of(context)
                 .primaryTextTheme
-                .headline2
-                .copyWith(fontWeight: FontWeight.w200),
+                .displayMedium
+                ?.copyWith(fontWeight: FontWeight.w200),
           ),
         ),
         Text(
-          "\$$sumMoneyAll",
+          '\$$sumMoneyAll',
           maxLines: 1,
-          style: Theme.of(context).primaryTextTheme.headline1,
+          style: Theme.of(context).primaryTextTheme.displayLarge,
         ),
       ],
     );
   }
 
   double get sumMoneyAll => currentProductList.fold(
-      0,
-      (previousValue, element) =>
-          previousValue + (element.price * element.count));
+        0,
+        (previousValue, element) =>
+            previousValue + ((element.price ?? 0) * (element.count ?? 0)),
+      );
 
   Expanded buildExpandedShopList() => Expanded(
-      flex: 4,
-      child: currentProductList.isEmpty
-          ? notFoundLottie
-          : ListView.builder(
-              itemCount: currentProductList.length,
-              shrinkWrap: true,
-              itemBuilder: (context, index) =>
-                  ShopPaymentCard(item: currentProductList[index]),
-            ));
+        flex: 4,
+        child: currentProductList.isEmpty
+            ? notFoundLottie
+            : ListView.builder(
+                itemCount: currentProductList.length,
+                shrinkWrap: true,
+                itemBuilder: (context, index) =>
+                    ShopPaymentCard(item: currentProductList[index]),
+              ),
+      );
 
   Widget get notFoundLottie =>
       Center(child: AspectRatio(aspectRatio: 4 / 5, child: NotFoundLottie()));
