@@ -4,15 +4,15 @@ import 'package:shoppifront/core/service/shop_dio.dart';
 import '../model/product.dart';
 
 class ProductService extends BaseService {
-  Future<List<ProductModel>> fetchProducts() async {
-    final response = await service.make<ProductModel>("product",
-        method: MethodType.GET, parserModel: ProductModel());
-    return response;
+  Future<List<Product>> fetchProducts() async {
+    final response = await service.make<Product>("product",
+        method: MethodType.GET, parserModel: Product());
+    return response is List<Product> ? response : [];
   }
 
-  Future postProduct(ProductModel model) async {
+  Future postProduct(Product model) async {
     final response = await service.make(routePath.product,
-        parserModel: ProductModel(), data: model, method: MethodType.POST);
+        parserModel: Product(), data: model, method: MethodType.POST);
     return response;
   }
 }
