@@ -4,8 +4,11 @@ import 'package:shoppingapp/features/shop/model/product.dart';
 
 class ProductService extends BaseService {
   Future<List<Product>> fetchProducts() async {
-    final response = await service.make<Product>(routePath.product,
-        parserModel: Product(), method: MethodType.GET);
-    return response;
+    final response = await service.make<Product>(
+      routePath.product,
+      parserModel: Product(),
+      method: MethodType.GET,
+    );
+    return response is List<Product> ? response : <Product>[];
   }
 }
