@@ -1,12 +1,11 @@
 import 'package:flutter/cupertino.dart';
 import 'package:provider/provider.dart';
+import 'package:shoppingapp/core/base/base_view_model.dart';
+import 'package:shoppingapp/core/constants/app_constants.dart';
 import 'package:shoppingapp/features/notifier/product_list_notifier.dart';
+import 'package:shoppingapp/features/shop/model/product.dart';
+import 'package:shoppingapp/features/shop/service/product_service.dart';
 import 'package:socket_io_client/socket_io_client.dart' as IO;
-
-import '../../../core/base/base_view_model.dart';
-import '../../../core/constants/app_constants.dart';
-import '../model/product.dart';
-import '../service/product_service.dart';
 
 class ShopViewModel extends BaseViewModel with ChangeNotifier {
   // IO.Socket socket = IO.io(AppConstants.SOCKET_URL);
@@ -35,7 +34,8 @@ class ShopViewModel extends BaseViewModel with ChangeNotifier {
   void initSocket() {
     socket.connect().on(AppConstants.SOCKET_CHANNEL_DELIVERY, (data) {
       // var jsonModel = jsonDecode(data);
-      var model = Product.fromJson(data);
+      if (data is! Map<String, dynamic>) return;
+      final model = Product.fromJson(data);
       final index = productList
           .indexWhere((element) => element.sId == data[AppConstants.S_ID]);
       if (index != -1) {
